+++ packages/scroll-mode/src/SwitchScrollModeDecorator.tsx
@@ -7,12 +7,10 @@
  */
 
 import * as React from 'react';
-import { LocalizationContext } from '@react-pdf-viewer/core';
+import { LocalizationContext, ScrollMode } from '@react-pdf-viewer/core';
 
 import { HorizontalScrollingIcon } from './HorizontalScrollingIcon';
-import { ScrollMode } from './structs/ScrollMode';
 import { VerticalScrollingIcon } from './VerticalScrollingIcon';
-import { WrappedScrollingIcon } from './WrappedScrollingIcon';
 
 interface RenderChildren {
     icon: React.ReactElement;
@@ -34,11 +32,6 @@ export const SwitchScrollModeDecorator: React.FC<{
         case ScrollMode.Horizontal:
             label = (l10n && l10n.scrollMode ? l10n.scrollMode.horizontalScrolling : 'Horizontal scrolling') as string;
             icon = <HorizontalScrollingIcon />;
-            break;
-
-        case ScrollMode.Wrapped:
-            label = (l10n && l10n.scrollMode ? l10n.scrollMode.wrappedScrolling : 'Wrapped scrolling') as string;
-            icon = <WrappedScrollingIcon />;
             break;
 
         case ScrollMode.Vertical: